+++ zimi/server.py
@@ -86,7 +86,7 @@ except ImportError:
 # SSL context using certifi CA bundle (PyInstaller bundles lack system certs)
 SSL_CTX = ssl.create_default_context(cafile=certifi.where())
 
-ZIMI_VERSION = "1.6.3"
+ZIMI_VERSION = "1.7.0"
 
 log = logging.getLogger("zimi")
 logging.basicConfig(